requirements.txt
fastapi
uvicorn[standard]
google-cloud-firestore
google-auth
requests
